Moved IModelIdentifierLister to Abstractions folder
