+++ src/Skyhop.Mail/Internal/ModelIdentifierLister.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Mvc.ApplicationParts;
 using Microsoft.AspNetCore.Mvc.Razor;
 using Microsoft.AspNetCore.Mvc.Razor.Compilation;
 using Microsoft.AspNetCore.Razor.Hosting;
+using Skyhop.Mail.Abstractions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
