+++ src/Skyhop.Mail/RazorViewToStringRenderer.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Mvc.ViewEngines;
 using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
+using Skyhop.Mail.Abstractions;
 using System;
 using System.IO;
 using System.Linq;
